+++ src/rutaAerea.java
@@ -1,5 +1,5 @@
 public class rutaAerea extends Ruta {
-    public rutaAerea(Terminal origen, Terminal destino, float distancia, float consumo, float tarifa, float tiempo) {
+    public rutaAerea(Terminal origen, Terminal destino, float distancia, float consumo, double tarifa, double tiempo) {
         this.origen = origen;
         this.destino = destino;
         this.distancia = distancia;
